pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eumdac-fetch"
version = "1.0.0"
description = "Bulk EUMETSAT data downloader"
readme = "README.md"
requires-python = ">=3.14"
license = "MIT"
authors = [{ name = "digital-idiot" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]
dependencies = [
    "eumdac",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "fsspec>=2023.0.0",
    "aiohttp>=3.8.0",
]

[project.urls]
Homepage = "https://github.com/digital-idiot/eumdac-fetch"
Repository = "https://github.com/digital-idiot/eumdac-fetch"
Documentation = "https://github.com/digital-idiot/eumdac-fetch#readme"
Issues = "https://github.com/digital-idiot/eumdac-fetch/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
]
docs = [
    "sphinx>=7.0",
    "furo",
    "myst-parser",
    "sphinx-copybutton",
    "sphinxcontrib-mermaid",
    "sphinx-click",
]

[project.scripts]
eumdac-fetch = "eumdac_fetch.cli:cli"

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ARG001", "ARG002"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--tb=short"
markers = [
    "integration: real API tests requiring EUMDAC credentials (use --run-integration)",
]
